Moving to Microservices - separating services into subfolders for monorepo approach. History, Search and User done, merged Meewat and fixed uploading issues, added new npm script to run both concurrently, adjusted webpack
