+++ HistoryService/client/components/simple_article.jsx
@@ -3,7 +3,7 @@
 */
 import React from "react";
 import { Link } from "react-router";
-import Loader from "../helpers/loader.jsx";
+import Loader from "../../../client/components/helpers/loader.jsx";
 import StatusAlert, { StatusAlertService } from "react-status-alert";
 
 class SimpleArticle extends React.Component {
