+++ SearchService/client/components/results.jsx
@@ -4,8 +4,8 @@
 import React from "react";
 import ArtworkResults from "./artwork_results.jsx";
 import Filters from "./filters.jsx";
-import BrowseTopics from "../wiki/browse_topics.jsx";
-import Loader from "../helpers/loader.jsx";
+import BrowseTopics from "../../../client/components/wiki/browse_topics.jsx";
+import Loader from "../../../client/components/helpers/loader.jsx";
 
 class Results extends React.Component {
   constructor(props) {
