+++ client/components/routes.jsx
@@ -11,20 +11,20 @@ import { BrowserRouter as Router, Switch } from "react-router-dom";
 // Import all components for navigation
 import App from "./main/app.jsx";
 import Home from "./main/home.jsx";
-import Login from "./users/login.jsx";
+import Login from "../../UserService/client/components/login.jsx";
 import Article from "./wiki/article.jsx";
 import NewArticle from "./wiki/new.jsx";
 import EditArticle from "./wiki/edit.jsx";
-import ArticleHistory from "./wiki/history.jsx";
-import Search from "./search/search.jsx";
-import Admin from "./users/admin.jsx";
-import Setup from "./users/setup.jsx";
-import EditTopic from "./users/edit_topics.jsx";
-import EditUser from "./users/edit_users.jsx";
-import UserSignup from "./users/user_signup.jsx";
+import ArticleHistory from "../../HistoryService/client/components/history.jsx";
+import Search from "../../SearchService/client/components/search.jsx";
+import Admin from "../../UserService/client/components/admin.jsx";
+import Setup from "../../UserService/client/components/setup.jsx";
+import EditTopic from "../../UserService/client/components/edit_topics.jsx";
+import EditUser from "../../UserService/client/components/edit_users.jsx";
+import UserSignup from "../../UserService/client/components/user_signup.jsx";
 import Landing from "./main/landing.jsx";
 import About from "./main/about.jsx";
-import Results from "./search/results.jsx";
+import Results from "../../SearchService/client/components/results.jsx";
 
 // --------------------------------------------------------------------------------------------------------------------------------------------
 // Build a route path and limit access to certain pages
